requirements.txt
selenium>=4.1.0
psutil>=5.9.0
python-dotenv>=0.19.2
